// extensions/3adel/src/models.d.ts
export interface DirectusAccess {

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   id: string;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   policy: DirectusPolicy | DirectusPolicy["id"];

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   role: DirectusRole | DirectusRole["id"] | null;

  /**
   * No description.
   *
   * Type in directus: integer
   * Type in database: integer
   */
   sort: number | null;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   user: DirectusUser | DirectusUser["id"] | null;
}

export interface DirectusActivity {

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   action: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   collection: string;

  /**
   * No description.
   *
   * Type in directus: integer
   * Type in database: integer
   */
   id: number;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   ip: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   item: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   origin: string | null;

  /**
   * No description.
   *
   * Type in directus: alias
   * Type in database: no column
   */
   revisions: DirectusRevision[] | null;

  /**
   * No description.
   *
   * Type in directus: timestamp
   * Type in database: timestamp with time zone
   */
   timestamp: string;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   user: DirectusUser | string | null;

  /**
   * No description.
   *
   * Type in directus: text
   * Type in database: text
   */
   user_agent: string | null;
}

export interface DirectusCollection {

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   accountability: string | null;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   archive_app_filter: boolean;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   archive_field: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   archive_value: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   collapse: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   collection: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   color: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   display_template: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   group: DirectusCollection | DirectusCollection["collection"] | null;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   hidden: boolean;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   icon: string | null;

  /**
   * No description.
   *
   * Type in directus: json
   * Type in database: json
   */
   item_duplication_fields: any | null;

  /**
   * No description.
   *
   * Type in directus: text
   * Type in database: text
   */
   note: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   preview_url: string | null;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   singleton: boolean;

  /**
   * No description.
   *
   * Type in directus: integer
   * Type in database: integer
   */
   sort: number | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   sort_field: string | null;

  /**
   * No description.
   *
   * Type in directus: json
   * Type in database: json
   */
   translations: any | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   unarchive_value: string | null;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   versioning: boolean;
}

export interface DirectusComment {

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   collection: DirectusCollection | string;

  /**
   * No description.
   *
   * Type in directus: text
   * Type in database: text
   */
   comment: string;

  /**
   * No description.
   *
   * Type in directus: timestamp
   * Type in database: timestamp with time zone
   */
   date_created: string | null;

  /**
   * No description.
   *
   * Type in directus: timestamp
   * Type in database: timestamp with time zone
   */
   date_updated: string | null;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   id: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   item: string;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   user_created: DirectusUser | DirectusUser["id"] | null;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   user_updated: DirectusUser | DirectusUser["id"] | null;
}

export interface DirectusDashboard {

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   color: string | null;

  /**
   * No description.
   *
   * Type in directus: timestamp
   * Type in database: timestamp with time zone
   */
   date_created: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   icon: string;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   id: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   name: string;

  /**
   * No description.
   *
   * Type in directus: text
   * Type in database: text
   */
   note: string | null;

  /**
   * No description.
   *
   * Type in directus: alias
   * Type in database: no column
   */
   panels: DirectusPanel[] | null;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   user_created: DirectusUser | DirectusUser["id"] | null;
}

export interface DirectusExtension {

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   bundle: string | null;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   enabled: boolean;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   folder: string;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   id: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   source: string;
}

export interface DirectusField {

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   collection: DirectusCollection | string;

  /**
   * No description.
   *
   * Type in directus: json
   * Type in database: json
   */
   conditions: any | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   display: string | null;

  /**
   * No description.
   *
   * Type in directus: json
   * Type in database: json
   */
   display_options: any | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   field: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   group: DirectusField | string | null;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   hidden: boolean;

  /**
   * No description.
   *
   * Type in directus: integer
   * Type in database: integer
   */
   id: number;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   interface: string | null;

  /**
   * No description.
   *
   * Type in directus: text
   * Type in database: text
   */
   note: string | null;

  /**
   * No description.
   *
   * Type in directus: json
   * Type in database: json
   */
   options: any | null;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   readonly: boolean;

  /**
   * No description.
   *
   * Type in directus: boolean
   * Type in database: boolean
   */
   required: boolean | null;

  /**
   * No description.
   *
   * Type in directus: integer
   * Type in database: integer
   */
   sort: number | null;

  /**
   * No description.
   *
   * Type in directus: csv
   * Type in database: character varying
   */
   special: string[] | null;

  /**
   * No description.
   *
   * Type in directus: json
   * Type in database: json
   */
   translations: any | null;

  /**
   * No description.
   *
   * Type in directus: json
   * Type in database: json
   */
   validation: any | null;

  /**
   * No description.
   *
   * Type in directus: text
   * Type in database: text
   */
   validation_message: string | null;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   width: string | null;
}
// ...
export interface DirectusFolder {

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   id: string;

  /**
   * No description.
   *
   * Type in directus: string
   * Type in database: character varying
   */
   name: string;

  /**
   * No description.
   *
   * Type in directus: uuid
   * Type in database: uuid
   */
   parent: DirectusFolder | DirectusFolder["id"] | null;
}
// ...
export type Collections = {
  directus_access: DirectusAccess[];
  directus_activity: DirectusActivity[];
  directus_collections: DirectusCollection[];
  directus_comments: DirectusComment[];
  directus_dashboards: DirectusDashboard[];
  directus_extensions: DirectusExtension[];
  directus_fields: DirectusField[];
  directus_files: DirectusFile[];
  directus_flows: DirectusFlow[];
  directus_folders: DirectusFolder[];
  directus_migrations: DirectusMigration[];
  directus_notifications: DirectusNotification[];
  directus_operations: DirectusOperation[];
  directus_panels: DirectusPanel[];
  directus_permissions: DirectusPermission[];
  directus_policies: DirectusPolicy[];
  directus_presets: DirectusPreset[];
  directus_relations: DirectusRelation[];
  directus_revisions: DirectusRevision[];
  directus_roles: DirectusRole[];
  directus_sessions: DirectusSession[];
  directus_settings: DirectusSetting;
  directus_shares: DirectusShare[];
  directus_translations: DirectusTranslation[];
  directus_users: DirectusUser[];
  directus_versions: DirectusVersion[];
  directus_webhooks: DirectusWebhook[];
}


export type CollectionName = keyof Collections;